Adicionando correções do Header e do Footer Co-authored-by: ArthurGuilher62 <Arthurguilher062@gmail.com>
