+++ client/src/components/Header/Header.css
@@ -1,78 +1,108 @@
 header {
-    background-color: #116FBB;
-    display: flex;
-    align-items: center;
-    column-gap: 32px;  
-    width: 100%;
-    padding-top: 1px;
-    padding-bottom: 2px;
-    flex-direction: row;
-    justify-content: center; 
+  background-color: #116FBB;
+  display: flex;
+  align-items: center;
+  column-gap: 2rem;  
+  width: 100%;
+  padding-top: 0.0625rem;  
+  padding-bottom: 0.125rem; 
+  flex-direction: row;
+  justify-content: center;
+  display: flex;
+}
+
+@media (max-width: 860px) {
+  header {
+    flex-direction: column; 
+    align-items: center;  
+    padding: 1rem 0; 
   }
-  
-  #title {
-    font-family: 'Season_Prime', sans-serif;
-    font-size: 46px;
-    color: white;
-    margin: 0;
-  }
-  
+
   .navigation {
     display: flex;
-    align-items: center;
-    column-gap: 48px;
-    list-style: none;
-    color: white;
+    flex-direction: column;  
+    row-gap: 1rem; 
+    align-items: center; 
+    width: 100%;  
+    padding: 0;
   }
-  
-  .navigation a {
-    text-decoration: none;
-  }
-  
+
   .navigation__link {
-    color: white;
-    font-size: 14px;
-    font-weight: bold;  /* Altere para 'bold' ou '700' para deixar o texto em negrito */
-    letter-spacing: 2px;
-    text-transform: uppercase;
-    position: relative;
+    font-size: 1rem;  
+    letter-spacing: normal;  
   }
-  
-  
-  .navigation__link::after {
-    content: '';
-    position: absolute;
-    left: 0;
-    bottom: -2px;
-    width: 70%;
-    height: 1px;
-    background: white;
-    transform: scaleX(0);
-    transform-origin: right;
-    transition: transform 0.6s;
+
+  .App-logo, .unb {
+    max-width: 4.375rem;  
   }
-  
-  .navigation__link:hover::after {
-    transform: scaleX(1);
-    transform-origin: left;
-    transition: transform 0.6s;
+
+  #title {
+    font-size: 1.5rem;  
+    margin-bottom: 1rem;  
   }
-  
-  .App-logo {
-    max-width: 90px;
-    margin-bottom: 0;
-  }
-  
-  .unb {
-    max-width: 90px;
-    margin-bottom: 0;
-    display: block;
-    margin-left: auto;
-    margin-right: auto;
-  }
-  
-  .logo-container {
-    display: flex;
-    align-items: center;
-  }
-  
+}
+
+#title {
+  font-family: 'Season_Prime', sans-serif;
+  font-size: 2rem;  
+  color: white;
+  margin: 0;
+}
+
+.navigation {
+  display: flex;
+  align-items: center;
+  column-gap: 3rem;  
+  list-style: none;
+  color: white;
+}
+
+.navigation a {
+  text-decoration: none;
+}
+
+.navigation__link {
+  color: white;
+  font-size: 0.875rem;  
+  font-weight: bold;  
+  letter-spacing: 0.125rem; 
+  text-transform: uppercase;
+  position: relative;
+}
+
+.navigation__link::after {
+  content: '';
+  position: absolute;
+  left: 0;
+  bottom: -0.125rem;  
+  width: 70%;
+  height: 0.0625rem; 
+  background: white;
+  transform: scaleX(0);
+  transform-origin: right;
+  transition: transform 0.6s;
+}
+
+.navigation__link:hover::after {
+  transform: scaleX(1);
+  transform-origin: left;
+  transition: transform 0.6s;
+}
+
+.App-logo {
+  max-width: 5.625rem;  
+  margin-bottom: 0;
+}
+
+.unb {
+  max-width: 5.625rem;  
+  margin-bottom: 0;
+  display: block;
+  margin-left: auto;
+  margin-right: auto;
+}
+
+.logo-container {
+  display: flex;
+  align-items: center;
+}
